refactor: move types to their own folder to simplify overall structure
