+++ tsconfig.json
@@ -16,5 +16,5 @@
     "strict": true,
     "target": "es2015"
   },
-  "include": ["./typings/**/*", "./src/**/*", "./test/**/*"]
+  "include": ["./src/**/*"]
 }
